Increase number of inodes in disk image.
By default, makefs allocates just enough inodes to
to give to the files and directories in the image, plus
a little extra slack.  This is much less than what newfs allocates.
Try to increase the number of inodes in the disk image, so that
when we run tests that create lots of files, we don't run out of inodes.