riscv64-test: use qemu's built-in OpenSBI
Since version 4.1.0, QEMU has shipped with a copy of OpenSBI as the
default firmware for the RISC-V platform, which can be selected with the
'-bios default' command line option.

Since OpenSBI is still in the early stages of development, breaking
changes to the firmware installation paths will occur, as is the case
for the upcoming v0.7 update. Switch to using the 'default' version of
the firmware for our QEMU tests to avoid being disrupted by these
breakages. This version lags behind the version in ports, but it should
be fine for the purposes of CI.