Unify usage of ARTIFACT_SERVER.
There were many hardcoded instances of artifact.ci.freebsd.org in the job scripts, and
a mix of places where it was assumed to have a https:// prefixed to it.

Unify on it being the bare hostname.

This allows my local cluster to be able to contact my local artifact server.