Load pflog and if_stf modules
There are new tests which require the use of the pflog and if_stf
modules. Load them so these tests are executed.

Note that the relevant tests have not yet been merged to stable/12, but
as they will be that change is included in this commit.

Sponsored by:   Rubicon Communications, LLC ("Netgate")