Remove "alltrace; dump;" from the ddb script in vm
alltrace is not very helpful in the console and mail because it contains
too much information but we only show last 200 lines in the mail.

dump is not kept and available after the test currently so it's no use
to generate it. The work of this part is still in progress.

Sponsored by:	The FreeBSD Foundation