Install sysutils/porch in the test image
As of FreeBSD commit
096c39fae4a ("tests: kern: add some porch(1)-based tty tests"), we have
some tty tests that will run if porch is installed.  Add it to the
test image now.

The commit in question hasn't yet been MFC'd to stable/13 and thus, the
test image for 13 does not *currently* need porch, but it's a pretty
light package and I do anticipate MFC'ing the tests in that direction.